backend/requirements-dev.txt
# Development dependencies for the eval baseline
# Install with: pip install -r requirements-dev.txt
pytest>=7.4
pyyaml>=6.0
